[DSPlayer] merged from trunk revs r31093:31242 (part 1)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/dsplayer@31244 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/dvdplayer/Codecs/ffmpeg/build_xbmc_win32.sh
@@ -1,10 +1,12 @@
 #!/bin/bash 
+if [ -d .libs ]
+then
 rm -r .libs
+fi
+
 make distclean
 
-./configure \
---extra-cflags="-fno-common -I../libfaad2/include -Iinclude/dxva2" \
---extra-ldflags="-L../../../../../system/players/dvdplayer" \
+OPTIONS="
 --enable-shared \
 --enable-memalign-hack \
 --enable-gpl \
@@ -20,7 +22,15 @@ make distclean
 --enable-muxer=spdif \
 --enable-muxer=adts \
 --enable-encoder=ac3 \
---enable-encoder=aac &&
+--enable-encoder=aac"
+
+if [ -f ../libvpx/libvpx.a ] 
+then
+echo Building with libvpx
+OPTIONS="$OPTIONS --enable-libvpx"
+fi
+
+./configure --extra-cflags="-fno-common -I../libfaad2/include -Iinclude/dxva2 -I../libvpx/" --extra-ldflags="-L../../../../../system/players/dvdplayer -L../libvpx" ${OPTIONS} &&
  
 make -j3 && 
 mkdir .libs &&
